Fix php cs
